Add files via upload
I am in the process of making UI edits for the ClientBridge- turned ContentBridge Dashboard view. I had originally created this for our original vision for our midterm. However, we chose to pivot and are now creating ContentBridge, a simplified version of our app so we can actually complete it. Our functionalities will look much different now- so our buttons have been updated. Our logo is in process still because we changed the name. I’m excited to see how we’ll integrate this into our app. I’ve only worked on this for about an hour this past week.