Add files via upload
RandomPhoto:
This Swift app, named "RandomPhoto," is a simple iOS application that fetches a random photo from the Unsplash API and displays it within an image view. The app consists of a single view controller with a pink background. Upon launching, it displays an image view and a button labeled "RandomPhoto." When the button is tapped, it triggers a method to fetch a new random photo from the Unsplash API and updates the image view with the fetched image. The image view is configured to fill its content with the aspect ratio preserved, and the button is positioned at the bottom of the screen. The app utilizes UIKit for the user interface components and networking to fetch the images.

This was a challenge that took me about 3 days to finish because I used UIKit rather than SwiftUI, which was new and just different. I wanted to mimic kind of what we did in the Cats or Dogs app we created in class