Add files via upload
BedtimeStory:
A few weeks ago, I created a small, visually appealing mock-up for a bedtime story app using SwiftUI, Apple's modern framework for building user interfaces. The goal was to design a simple yet engaging user interface that could serve as a part of a larger app dedicated to bedtime stories. I did not focus on functionality for this app, I really just wanted to play around with the UI!
Code Overview:
The project consists of two main views:
ContentView:
This is the primary view where users first interact with the app. It presents a list of bedtime stories, each accompanied by a star icon to add a playful element to the display.
The list is implemented using SwiftUI's NavigationView and List, ensuring a native look and feel that's both performant and smooth.
Each story is represented by a Story struct that contains a title and a summary. This struct makes it easy to handle and display story data throughout the app.
The styling of the list items and the navigation bar is customized to fit the bedtime theme, with dark backgrounds and contrasting text colors to enhance readability.
StoryDetailView:
This view displays the details of a selected story. It's designed to be both colorful and simple, enhancing the reading experience without distracting from the content.
The view uses a VStack to vertically align the text elements, which include the story’s title and summary. The text is styled with different fonts and colors (white for the title and yellow for the summary) to differentiate the elements visually.
The background of the view features a linear gradient that transitions from black to blue, giving it a dreamy and pleasant aesthetic that fits the bedtime story theme.
The layout ensures that the text is both aesthetically pleasing and easy to read, with generous spacing and padding.I don’t think it looks as great as it could, but I did my best on this view.